+++ backend/app.js
@@ -1,28 +1,19 @@
 const express = require("express");
-const https = require("https");
 const cors = require("cors");
-const { Server } = require("socket.io");
-const { responseHandler } = require("./middleware/responseHandler");
 const path = require("path");
-const httpsOptions = require("./config/httpsOptions");
-const chatService = require("./services/chatService");
+const { responseHandler } = require("./middleware/responseHandler");
 const authRoutes = require("./routes/authRoutes");
 const streamRoutes = require("./routes/streamRoutes");
 const chatRoutes = require("./routes/chatRoutes");
-const { startFfmpegStream } = require("./services/streamService");
 
-// Start Ffmpeg Stream
-startFfmpegStream();
-
-// Setup Express
 const app = express();
+
+// Middleware
 app.use(cors());
 app.use(express.json());
-
-// Response Middleware
 app.use(responseHandler);
 
-// Setup routes
+// Routes
 app.use("/api/auth", authRoutes);
 app.use("/api/stream", streamRoutes);
 app.use("/api/chat", chatRoutes);
@@ -34,12 +25,4 @@ app.get("/", (req, res) => {
 });
 app.use("/stream", express.static(path.join(__dirname, "stream")));
 
-// Start Chat Server
-const server = https.createServer(httpsOptions, app);
-const io = new Server(server);
-chatService(io);
-
-// Start HTTPS server
-server.listen(8000, () => {
-  console.log("Server running at https://localhost:8000");
-});
+module.exports = app;
